[FrameworkBundle] Allow to use the BrowserKit assertions with Panther and API Platform's test client
