+++ src/Symfony/Component/DomCrawler/Test/Constraint/CrawlerSelectorAttributeValueSame.php
@@ -47,7 +47,7 @@ final class CrawlerSelectorAttributeValueSame extends Constraint
             return false;
         }
 
-        return $this->expectedText === trim($crawler->getNode(0)->getAttribute($this->attribute));
+        return $this->expectedText === trim($crawler->attr($this->attribute));
     }
 
     /**
